backend/requirements.txt

fastapi
uvicorn
azure-identity
azure-mgmt-resource
azure-mgmt-consumption
pandas
